Extract app.py CLI dispatch and port-preflight into vtsearch modules
Move the ~625-line __main__ argparse/autodetect block to
vtsearch/cli_main.py (decomposed into _build_parser + per-concern
helpers so each stays under the McCabe gate) and the Linux /proc
port-preflight + single-instance lock to vtsearch/port_preflight.py.
app.py's __main__ now delegates to cli_main.main(app, initialize_server),
passing the WSGI app and initialize_server in rather than re-importing
app under a second module name. app.py drops from 1462 to ~630 lines,
keeping only the WSGI app object, request hooks, error handlers,
blueprint registration, and initialize_server.
