+++ tests/core/test_startup_port_preflight.py
@@ -1,9 +1,12 @@
-"""Tests for the startup port-collision preflight helpers in ``app.py``.
+"""Tests for the startup port-collision preflight helpers.
 
 These cover the detection + safe-exit paths (``_port_is_free``,
 ``_find_listener_pids``, the non-TTY bail) and the real SIGTERM/SIGKILL
 ``_terminate_listeners`` path against a throwaway child process. They use
 ephemeral ports so they never collide with a real instance on :5000.
+
+The helpers live in ``vtsearch.port_preflight`` (extracted from ``app.py``);
+only ``python app.py``'s ``__main__`` launch path calls them.
 """
 
 import io
@@ -15,7 +18,7 @@ import time
 
 import pytest
 
-import app as app_module
+import vtsearch.port_preflight as app_module
 
 
 def _listening_socket():
